Rename kalman_localization.py -> kalman.py
