+++ lib/__init__.py
@@ -5,4 +5,4 @@
 """
 
 from lib.localization import *
-from lib.kalman_localization import *
+from lib.kalman import *
